arm: highmem: Add support for flushing kmap_atomic mappings
This is a snapshot from msm-4.4 branch of the
commit 9c2442a7a957 ("arm: highmem: Add support for flushing
kmap_atomic mappings").

This patch implements highmem code which provides kmap_flush_unused
to ensure all kmap mappings are really removed if they are unused.
This code does not handle kmap_atomic mappings since they are managed
separately. This prevents an issue for any code which relies on
having absolutely no mappings for a particular page. Rather than pay
the penalty of having CONFIG_DEBUG_HIGHMEM on all the time, add
functionality to remove the kmap_atomic mappings in a similar way to
kmap_flush_unused.

Change-Id: If324e6c96250a894cfdb0876e0b096de1e3efca8
Signed-off-by: speriaka <speriaka@codeaurora.org>