msm: make pmu device independent from perf framework
pmu.c is calling a function defined in perf_event.c. This function may
or may not be defined depending on whether PERF was enabled in the kernel
config
As the point of using the enable_irq_callback/disable_irq_callback functions
from the perf file was to enable/disable irqs on one particular CPU, it's
really not related to perf at all, therefore this change creates an equivalent
function locally to do the same operation.

Change-Id: I19378de406ef6756ee368bd44ea75d656acbb4a1
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>