msm: ipq806x: remove mention of cdp in init function
Current init function is called "ipq806x_cdp_init()" but the
IPQ806x product does not have a CDP board.

Instead of a single CDP board, IPQ uses a common init
function for all boards and then has the driver test the current
platform data to see if the driver should be enabled.

This change updates the "init_machine" function name to reflect
the fact that it is not tightly coupled to a particular board
but is generic for all ipq806x based platforms.

Change-Id: Ie1a78ba08699e1f0892ffd7b4fc006ed8746f0a0
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Sol Kavy <skavy@codeaurora.org>