Revert "net: socket ioctl to reset connections matching local address"
This reverts commit 2365d054980471b7fcb6d862169ddcec1eb52f4d.

As tcp.c is using "#if defined(CONFIG_IPV6_MODULE)" to detect
whether IPV6 is enabled or not, and as tcp.c is built
independently of this option, enabling IPV6 in openwrt results
in build errors: undefined reference to `rt6_lookup' and
`in6addr_any'.

Indeed, in our case tcp.c is linked statically and IPV6 is built
as a kernel module.

Conflicts:
	net/ipv4/tcp.c

Change-Id: If721dff8744503ffffa99399a5cbab607b3f15a5
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>