crypto: ocf: remove the non-integrated hardware from the Makefile
Hardware not used in OCF has not been integrated in the kernel. But
the patching actually adds a bunch of hardware support and the
corresponding directory. It's not a problem at build time but the
kernel reports an error when running clean.

This change removes this directories from the build so clean
succeeds.

Change-Id: I38e2331abd18ed59984b97e2e5bbf87565985571
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>