msm: ipq806x: Update platform names to use existing format
Existing platform use a specific format to
declare the board. This change updates the current format to
reflect the format that was used so far.

These strings is used by init scripts to know which board
we run in case we need to do some board specific default
configuration (gmac, switch...).

Change-Id: Id0a0e0dbb93b509624bf64eccc94ce647ed56ff4
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>