Revert "USB: Add nop functions for !USB case"
This reverts commit 15f25535688609ec29b7f6e5921721953ef33345.

This commit added a bunch of functions added in the case where
USB OTG were selected but not USB. If CONFIG_USB is not
present it defines some stub functions.

This approach doesn't work when USB support is built as module.
In that case, CONFIG_USB will be null but the real function will
actually be defined, therefore causing an error like
"redefinition of 'usb_blahblah...', this function existed in ..."

The root problem here is at a Kconfig level, not in the code: if
OTG needs a function in the USB driver, it should depend on it
and we should not redefine a stub function. Besides, if it works
with a stub function that does nothing, we probably don't need to
call it anyway.

Anyway, looking at the Kconfig, it seems that the dependency is
there. So this part is deprecated; therefore we'll revert the
initial workaround.

Change-Id: Ie32e6d3bcac7a43a65e5eb25f812d6b72a1bc315
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>