drivers: usb: remove usb_[un]register_notify stub functions
These functions are defined conditionally on #ifdef CONFIG_USB,
which is a mistake. If we enable CONFIG_USB as a module (like
it's the case in QSDK), then these functions will be defined
twice.

The proper fix consists in adding a dependency at the Kconfig
level to make sure the right modules are enabled when they're
needed. As the only drivers who are using these functions are
in the USB directory, then we can safely assume these functions
will always be there. Therefore removing the stub functions.

Change-Id: I73a1c7858723ad83c59d4c959d0d11d06959f407
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-by: Sol Kavy <skavy@codeaurora.org>