Add bridge API to access the bridge slave port given a bridge and a MAC address
This API is needed by the NSS connection manager modules to enable NSS fast path for
bridge flows. The connection manager needs to inspect the netdevice of the bridge slave
port that has learned a given MAC address. Currently no such API exists that is
exported out of the kernel network stack for this purpose.

Change-Id: Ifd4f2a447768d17a8b7b4b862f6971e45a791791
Signed-off-by: Kiran Kumar C.S.K <kkumarcs@codeaurora.org>
Reviewed-by: Bob Amstadt <ramstadt@codeaurora.org>