mtd: msm_nand: Fix error return on ECC errors
If ECC errors occur, during a nand_read operation, the read should
read and return the length requested. But should return with -EBADMSG
or -EUCLEAN as return value. If the read does not read the entire
length UBI layer treats it as an IO error.

During improper shutdowns a partially written page triggers
uncorrectable ECC errors. If the nand_read is not implemented as
mentioned above, the EIO error will result in a unmountable
filesystem. If the nand_read is implemented correctly, the filesystem
will recover, and last written alone will be lost.

Change-Id: I35a6ad9ed1743fb7fadb933a90f319b067173016
CRs-Fixed: 549291
Signed-off-by: Vijay Kumar <c_vbagav@codeaurora.org>
Reviewed-by: Varadarajan Narayanan <varada@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>