[ipq806x] Disabling request of Idle mode frequency for APPS Fabric
When Krait runs at 384 MHz , it requests for Apps Fabric to be at 133 MHz,
but DDR runs at 400 MHz in this setting. Fabric at 133 MHz and DDR at 400 MHz is not a
hardware recommended setting. So we are fixing this by Kraits always requesting for Apps Fabric
of 400 or 533 MHz, at which speeds the DDR runs

CRs Fixed: 553779, 556706, 556687

Change-Id: Id8707588cd87200149f6927f924c53e186f07039
Signed-off-by: Pamidipati, Vijay <vpamidip@codeaurora.org>
Reviewed-by: Abhishek Rastogi <arastogi@codeaurora.org>