msm: Remove Kconfig options & files that are no longer used
The working branch consists of patches from Google that were
based on top of v3.4. On top of Google's patches, support for
the ARM-MSM platform has been added.

This patch aims at modifying the contributions for MSM by
removing code associated with Kconfig options that are no
longer used in IPQ806x.

Change-Id: I9581e05426c78fe19601bc1a09f8782f6ec150a8
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>