Revert "mmc: card: Add Sanitize unit test"
This reverts commit 22f7abfd7f866ce603735f00a833fec391aec877.

	mmc: card: Add Sanitize unit test

	This test write data to the card, then send DISCARD on random
	addresses on the card, send SANITIZE to the card to erase all
	the unmapped areas.

This patch is part of the test io scheduler support, which allows
testing a block device by dispatching specific requests according
to the test case and declare PASS/FAIL according to the requests
completion error code.

The test IO scheduler in not enabled in the akronite defconfig, so remove it.

Conflicts:

	drivers/mmc/card/mmc_block_test.c

Change-Id: I5bd94e8ab7cae57b63570f4d7620666f34753914
Signed-off-by: Georgi Djakov <gdjako@codeaurora.org>
Signed-off-by: Srinivas Dodda <doddas@codeaurora.org>