Revert "block: test-iosched error handling fixes"
This reverts commit db8cbfe2b32773a619c3cdbf8e99bbd7954a7d4f.

	block: test-iosched error handling fixes

	- Fix test-iosched crash when running multiple tests
	- Free the BIOs memory when a request is not completed

This patch is part of the test io scheduler support, which allows
testing a block device by dispatching specific requests according
to the test case and declare PASS/FAIL according to the requests
completion error code.

The test IO scheduler in not enabled in the akronite defconfig, so remove it.

Change-Id: Iab4208a316c4b76c95650fce3a811a9a1ccbdbf7
Signed-off-by: Georgi Djakov <gdjako@codeaurora.org>
Signed-off-by: Srinivas Dodda <doddas@codeaurora.org>