sound: ipq: Multiple file play option in Aplay
When multiple files are specified to aplay in command
line option, aplay does not close the interface after
playing a song. Instead it simply resets the HW params
and sends the data again. So we need to reinit all our
stuff in set HW params instead of interface open.
This way we will able to support multiple file option.

CRs-Fixed:  589844

Change-Id: I9c8af2f81658f271b21c57e309b3772a8c7c8bce
Signed-off-by: Anusha Canchi Ramachandra Rao <anusha@codeaurora.org>