sound: ipq: Improved interrupt and buffer management to prevent noise
Simplified Rx/Tx API and made it into one along with configuring
for only Rx DMA interrupts and not both. The read data pending is
now only one flag instead of accumulated count. The interrupt status
registers are cleared before enabling the interrupt. Added new API
for VOIP stack to indicate when the VOIP stack is done using the
buffer.

Change-Id: I749da769d3521a83fcca30c4dd4d874b206277d8
Signed-off-by: Saravanan Jaganathan <sjaganat@codeaurora.org>