msm: ipq806x: Generate fabric registers dump on L2 cache error
Currently the L2 cache error generates an interrupt and dumps
the L2 cache related registers. But these don't capture the
offending address (i.e. the address that caused the L2 cache error).
But, the offending address is captured by one of the five fabric registers.
As the exact fabric which captured the offending address is not
known, we dump the important registers of all the fabrics.

This can be enabled using the Kconfig option 'DEBUG_IPQ806X_CACHE_ERP_REGS'

Change-Id: I4229ab4acc033e7f7f068207e0f38edab015e1cd
CRs-Fixed: 584141
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>