ipv6: fix race condition regarding dst->expires and dst->from.
commit ecd9883724b78cc72ed92c98bcb1a46c764fff21 upstream.

This upstream commit was modified to move the XFRM pointer to the end
of the dst_entry structure. This was done to preserve the 64B alignment
of __refcnt, which is mandatory.
The upstream commit doesn't have this problem as it can afford removing
the *pad0 pointer from the structure introduced by the commit removing
the *_neighbour pointer. Unfortunately, the *_neighbour pointer cannot
be removed in our case as a lot of code still relies on it.

Eric Dumazet wrote:
| Some strange crashes happen in rt6_check_expired(), with access
| to random addresses.
|
| At first glance, it looks like the RTF_EXPIRES and
| stuff added in commit 1716a96101c49186b
| (ipv6: fix problem with expired dst cache)
| are racy : same dst could be manipulated at the same time
| on different cpus.
|
| At some point, our stack believes rt->dst.from contains a dst pointer,
| while its really a jiffie value (as rt->dst.expires shares the same area
| of memory)
|
| rt6_update_expires() should be fixed, or am I missing something ?
|
| CC Neil because of https://bugzilla.redhat.com/show_bug.cgi?id=892060

Because we do not have any locks for dst_entry, we cannot change
essential structure in the entry; e.g., we cannot change reference
to other entity.

To fix this issue, split 'from' and 'expires' field in dst_entry
out of union.  Once it is 'from' is assigned in the constructor,
keep the reference until the very last stage of the life time of
the object.

Of course, it is unsafe to change 'from', so make rt6_set_from simple
just for fresh entries.

CRs-Fixed: 583433
Change-Id: Ie35e7242f3c80e4a422fc34f91f4788ad33b32e9
Reported-by: Eric Dumazet <eric.dumazet@gmail.com>
Reported-by: Neil Horman <nhorman@tuxdriver.com>
CC: Gao Feng <gaofeng@cn.fujitsu.com>
Signed-off-by: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>
Reviewed-by: Eric Dumazet <edumazet@google.com>
Reported-by: Steinar H. Gunderson <sesse@google.com>
Reviewed-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>