msm: spi: Replace workqueue with kthread with RT priority
The voip stack has a duty cycle of 10 ms. One of the activities to
be completed during this period is accessing the SLIC registers
via SPI. The SPI driver currently uses a workqueue to handle the
SPI messages from the kernel. As this does not have a high priority,
this adds a few ms latency during wakeup after sleep.  So replacing
the workqueue with a high priority RT thread reduces these delays.

Also, spi_sync latency is reduced considerably if the SPI interrupt
and the handler thread is run on the same core. So we are adding
support to receive notifications when the IRQ affinity is changed
and to update the thread affinity.

Change-Id: I7f536acf09ebb85a8c4bf3482b4a2495498fac0e
CRs-Fixed: 619750
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>