net: ipv4: Fix build failure when IP GRE is enabled
A regression was introduced in commit
8a533666d1591cf4ea596c6bd710e2fe682cb56a

This commit was backported from upstream but some API have been removed
in the meantime. dst_get_neighbour() is no longer used. This went
undetected so far as we don't use IP GRE, but now that we want to enable
PPTP & L2TP, this causes a build failure.

The neigh field is set just below so the failing line is not required.
It doesn't exist in the upstream commit either so I assume it was an
integration mistake.

Change-Id: I2f424c07b1c4df72e1f3e4b3e05cf3575c53619c
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>