ASoC: Add macros for defining virtual enums
With the upcoming consolidation of normal MUXs and virtual MUXs we need to be
able to distinguish between enums with and without a backing register at the
enum level. Use the same approach as used for virtual mixer controls by setting
the reg field of the enum to SND_SOC_NOPM for enums without a backing register.
This patch adds a set of helper macros that can be used to define such enums.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>
 (cherry picked from commit b948837a32cc2a510c2baf90111c8c5dde801a66)

Change-Id: I6c35b4ce4b6f1e9fb780a8b412edae0c08f78d93
Signed-off-by: Ramprasad Katkam <katkam@codeaurora.org>