Revert "wext : Increase iw_priv_args->name array length by 16"
This reverts commit ab221d7c942b65c403ee76b8dc3e1c8bea1cb4ba.

This change prevents qca-wifi from running on a non-upstream kernel.
It assumes that every WiFi user will have the change in wireless.h
which is not a safe assumption. A lot of users do not want to modify
the standard mechanisms; this change should never have been submitted
at the beginning.

Change-Id: I3b3f6a2ac4f66f1c1609461b89e790ee947e7877
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>