[skbuff] Add Maximum Recycle Size
A bug from the field revealed an incompatibility with the gigantic 8K+
skbuffs used by the QCA wifi driver.  If the wifi driver was passing
traffic, AND a very large UDP (or other fragmented protocol) datagram
was sent, recycling these large buffers for this purpose would cause
the default value for ipv4.frags.high_thresh to be exceeded, and the
datagram will be dropped with a statistical frequency relative to the
percentage of wifi skbs to normal SKBs.

This does NOT fix the underlying issue of huge, multi-page skbs used
by the wifi driver.  It is likely that the above issue exists for the
wifi driver by itself, and it should be tested.

I am not very happy with the "width" of the recycleable SKBs, as this
greatly reduces the percentage of SKBs that can be recycled for some
drivers.  This will need to be improved in future iterations.

Thanks to Qu Wenyuan for identifying this issue, and getting to a root
cause.

Change-Id: I7c0aa115b09c49d2ba51515e2727ae14006f4d9a
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>