[skbuff] Reduce Max Recycle SKBs to 1024
This change reduces the maximum recycle SKBs to 1024 from 2048.

Due to the unusually large size of SKBs used by wifi, the per-core
cache in the recycle list can grow much larger than originally
intended.

This change reduces the total SKBs back to a more reasonable level.
Additionally, this change still leaves the recycle lists close to
the total number of SKBs required: the sum of the size of all of
the RX rings in the system.

This is a static configuration, which does not allow the system
administrator to control the degree to which SKBs are cached.  This
should be improved in future implementations.

Change-Id: I34c8f9e474b3b9873e88376b6dc49b8470d13da4
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>