OpenWRT: add a ubi forced detach ioctl
When running ubi rootfs upgrade on nand based OpenWrt device after
pivot_root init process is still "hooked" on the old file system.
Thus, the old file system can not be umounted. If the filesystem is
mounted it can not be upgraded with using for example ubiupdatevol
or removed with ubirmvol. Forcing umount would allow to run the
before mentioned commands.

Change-Id: I079febb7f547c8ac9fb733ae6c421001d983e7f9
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Signed-off-by: Sachin Sundar <ssundar@codeaurora.org>
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>