FROMLIST: dmaengine: qcom_bam_dma: Add v1.3.0 driver support
This patch adds support for the v1.3.0 version of the BAM dma ip block.  This
patch adds register access abstraction to deal with the changes to the register
map between the two versions.  Blocks of registers moved around within the
address space, and multipliers used for calculating the pipe registers changed
as well.

Signed-off-by: Andy Gross <agross@codeaurora.org>