[skbuff] Increase Max Recycle to Include 4K SKBs
The original patch to decrease the maximum recycle size to address
several memory-related issues reduced the maximum SKB that could be
recycled so that it did NOT include 4K-page SKBs.  Wifi uses 4K-page
SKBs regularly, and this caused wifi SKBs to not be freed to the
recycle list (failing the max size check).

This change increases the size to be certain to include 4K-page skbs.
The true number should be:
 4096 - SKB_DATA_ALIGN(sizeof(struct skb_shared_info)),
or approximately 3840B.  We are over-shooting here a bit for testing
purposes, and in production, this difference does not matter; the
underlying allocator only allocates kmalloc-pool-sized SKBs, we
simply must ensure we include 4K, and exclude 8K.

Thanks Aaditya for helping with the root cause of the performance
decrease on wifi due to the original limit.

Change-Id: Ia9f874c92ab695f983eb63140a1470230adab38e
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>