leds: usbdev: Fix BUG: scheduling while atomic
When 'ledtrig_usbdev' tries to turn on/off the LEDs corresponding
to USB device insertion/removal usbdev_trig_name_store seems to
acquire a write lock on td->lock and calls usb_find_device_by_name.

usb_find_device_by_name in turn, tries to acquire a mutex lock
(either on usb_bus_list_lock or bus->root_hub). In some rare
cases, if the lock is not available, it might go to sleep,
resulting in the 'scheduling while atomic' BUG assertion as
follows.

	BUG: scheduling while atomic: S96led/2428/0x00000002
	Modules linked in: ...
	Backtrace:
		(__schedule_bug+0x0/0x54)
		(__schedule+0x0/0x4fc)
		(schedule+0x0/0x70)
		(schedule_preempt_disabled+0x0/0x38)
		(__mutex_lock_slowpath+0x0/0x18c)
		(mutex_lock+0x0/0x48)
		(usb_find_device_by_name+0x0/0x84 [usbcore])
		(usbdev_trig_name_store+0x0/0xb4 [ledtrig_usbdev])
		(dev_attr_store+0x0/0x2c)
		(sysfs_write_file+0x0/0x16c)
		(vfs_write+0x0/0x13c)
		(sys_write+0x0/0x70)

To overcome this, take a copy of the data that is needed by
usb_find_device_by_name and invoke it with the copy instead of
the locked data structure.

CRs-Fixed: 652860

Change-Id: I7edeb0fbad0c3f013053614d0376bc6b01486473
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>