FROMLIST: msm: scm: Fix incorrect cache invalidation
The cache invalidation in scm_call() correctly rounds down the
start address to invalidate the beginning of the cacheline but
doesn't properly round up the 'end' address to make it aligned.
The last chunk of the buffer won't be invalidated when 'end' is
not cacheline size aligned so make sure to invalidate the last
few bytes in such situations. It also doesn't do anything about
outer caches so make sure to invalidate and flush those as well.

BUG=None
TEST=emerge-storm chromeos-kernel-3_14

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(am from https://patchwork.kernel.org/patch/4674401/)

Change-Id: I8784979fb1dcbbd8e49cc17201e5af3c81908100
Signed-off-by: Yogesh Lal <ylal@codeaurora.org>