FROMLIST: msm: scm: Add atomic SCM APIs
The atomic SCM APIs are useful for commands that are guaranteed
by the secure side to be uninterruptable, atomic and SMP safe.
The calling convention use registers for passing parameters and
return values between the secure and non-secure side. Support
this interface with  scm_call_atomic[1-2]() functions
corresponding to the number of arguments passed.

BUG=None
TEST=emerge-storm chromeos-kernel-3_14

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(am from https://patchwork.kernel.org/patch/4674271/)

Change-Id: I479bf3afd189c4d19dfc9c70ad0fb89d0787213f
Signed-off-by: Yogesh Lal <ylal@codeaurora.org>