sch_blackhole: return NET_XMIT_DROP on enqueue
return DROP instead of SUCCESS.  SUCCESS is interpreted as
a successful enqueue, but the blackhole qdisc enqueues nothing.

this becomes a bigger problem when sch_blackhole is paired
with sch_hfsc (and possibly others but the research stopped with
hfsc), which checks whether the enqueue to the hfsc class's attached
qdisc was successful, and if so increments internal state.  in this
case, hfsc increments its own qlen counter but since the skb
was actually dropped by the qdisc, hfsc's counters are incorrect,
leading to WARN_ONs in the watchdog schedule function and
eventually causing a system reboot.

Change-Id: If847e9663f3a6868a5accc005687dcab68373689
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>