[skb recycle] Set Max Recycle Size to 4K Page
This change sets the maximum recycle size for the SKB recycler to a
single 4K page.  This ensures that we will not suffer the performance
loss associated with zeroing and invalidating 8K pages, and serves as
a check that we are not allocating these larger pages.

If needed, we can add a separate recycler for 4K and 8K pages, but on
MIPS this produced a significant performance increase.

Change-Id: I964ffe4cc1285e19f0d07f7de4958f6ab3c9c5ce
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>