prng: Update msm-rng driver to wait for data availability
The msm-rng driver currently returns immediately if there is
no data available to be read from the PRNG block. This causes
the user space 'rngd' to continuously read from this
device (/dev/hwrng) and the CPU utilization goes up.

We dont see this problem when we run rngd to read from
/dev/urandom instead of /dev/hwrng

Two changes seem to fix this:
1. Keep the clock enabled (currently the clock is enabled during
read and then disabled)
2. Sleep for a short duration if not data is available from PRNG
and read again, this would cause reads on /dev/hwrng to block util
there is data available.

Change-Id: I8a6b5dca846f3a2c615e8dbcb513ab6a3e95387b
CRs-Fixed: 736422
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>