ARM: SMP: Add temporary support in absence of TZ
TZ holds the secondary cores till it is woken up by the
primary core. In the pre-silicon platforms when the TZ code
is not available, there has to be way to hold the secondary cores.
Adding the support here. This is not required when the TZ is
available.
The code added in files arch/arm/mach-qcom/headsmp.S and arch/arm/mach-qcom/platsmp.c
has been resued from the 3.10 kernel from the below repo
https://www.codeaurora.org/cgit/quic/qsdk/oss/kernel/linux-msm/log/?h=caf2/msm-3.10
Files from which the code is partially/fully resued are
arch/arm/mach-msm/headsmp.S
arch/arm/mach-msm/platsmp.c

Change-Id: I105b8370c85eb863836bf51357a2b173c6b816cd
Signed-off-by: Sricharan R <sricharan@codeaurora.org>