msm_serial_hs: Rx discard flush timeout
While configuring the uart buadrate using set_termios
function, timeout is occuring for rx discard flush
operation. This is due to the driver is waiting for
less time for the DMA call back to occur after programing
the DMA register with the discard flush command.

To overcome this issue increased the timeout to 300ms.

Change-Id: Ifbf807e5c28e9cbed986e7b4a45f31d628307d1f
Signed-off-by: Dilip Kota <c_dkota@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>
Signed-off-by: Nisar Ahamed Khan <nkhan@codeaurora.org>