msm_serial_hs: Programming Auto RFR Ctrl bit
1.) Disable the auto RFR before forcibly making the RFR
line HIGH from the driver to avoid the controller
changing it's state with out driver's knowledge and
enable the auto RFR instead of making the RFR line
forcibly LOW.So that controller will make it LOW
based on the rx FIFO condition
2.) Due to unexpected data received from the client during
clock off sequence, the clock off sequence is going
into wrong state and eventually leads to uart hardware
mess up.

This patch doesn't allow any data from the client during
clock off sequence is in progress

This is a snapshot of the MSM serial hs driver as of kk_2.7 commit.
https://www.codeaurora.org/cgit/quic/la/kernel/msm/commit/drivers/tty/serial/msm_serial_hs.c?h=kk_2.7&id=9fa49487e444a77b76291b073fd982c5875cc4da

CRs-Fixed: 546216

Change-Id: Ib6d3ca4eab797554197a3d7e2364ac3b8d155557
Signed-off-by: Bharath Gandhi <bgandhi@codeaurora.org>