CHROMIUM: cpufreq: krait: allow selection of OPP table based on chip characteristics
In order to properly scale the Kraits, it is necessary to choose an
appropriate table based on the characteristics of the chip.  Read from
the FUSE data via syscon and use it to select the correct table.

BUG=chrome-os-partner:32317
TEST=Verify CPU frequency scaling happens when load alters

Change-Id: I75a1c10d5822a9519b10d6679cb3939cdb979727
Signed-off-by: Josh Cartwright <joshc@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/231792
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Tested-by: Viswanath Kraleti <vkraleti@codeaurora.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>

Change-Id: I48bc07535e76ed1cba76a2e24890c172c493e929
Git-commit: a2979c3ae755b7c86533448791937cb2ee1c3814
Git-repo: https://chromium.googlesource.com/chromiumos/third_party/kernel
Signed-off-by: Varadarajan Narayanan <varada@codeaurora.org>
Signed-off-by: Rajith Cherian <rajith@codeaurora.org>