sched: Fix incorrect softirq accounting when IRQ time accounting is enabled
When ksoftirqd runs, cpu_sotfirq_time is not updated (to account
time to correctly to the ksoftirqd task), but
cpustat[CPUTIME_SOFTIRQ] continues to be incremented. So the
difference in the values of cpustat[CPUTIME_SOFTIRQ] and
cpu_softirq_time increases, and is propotional to the duration
ksoftirqd runs. After ksoftirqd sleeps, the cpu_softirq_time starts
being updated again, but because of the previous difference in
values it takes time to match the value in cpustat[CPUTIME_SOFTIRQ].
So during this period, this function returns 0 which causes incorrect
softirq values in /proc/stat. So the solution is to plug the
difference in values using another variable (cpu_ksoftirqd_time)
which is updated when ksofitrqd runs.

Change-Id: I08f80460b8470a04c1ec4ee08876c4b0eaef903e
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>
Signed-off-by: Kirthik Srinivasan <kirthik@codeaurora.org>
Signed-off-by: Jaiganesh Narayanan <njaigane@codeaurora.org>