Openwrt: Apply 120-bridge_allow_receiption_on_disabled_port.patch
When an ethernet device is enslaved to a bridge, and the bridge STP
detects loss of carrier (or operational state down), then normally
packet receiption is blocked.

This breaks control applications like WPA which maybe expecting to
receive packets to negotiate to bring link up. The bridge needs to
block forwarding packets from these disabled ports, but there is no
hard requirement to not allow local packet delivery.

Change-Id: Ifcb106549c4b724acca53be6cf0599cb91a657b4
Signed-off-by: Padma Raghunathan <padmar@codeaurora.org>