clk: qcom: ipq806x: Fix ubi32 core clock scaling
The ubi32 core clocks were getting turned off during late init,
but the ubi32 core clocks need to stay on because the processor
is running out of the bootloader. Mark the clocks as
IGNORE_UNUSED so that late init doesn't turn these clocks off.
This makes the bank switching during clk_set_rate() work
properly.

The next problem is the safe_parent op is incorrectly
indicating PLL18 instead of PLL8 as the safe parent to use during
clock scaling. Fix this by using the software index into the parent
array instead of the hardware index for PLL8 (which happens to
map to the PLL18 software index).

Change-Id: I83d55c8945575f309d58e74830a92dc502a524a5
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>