tty: serial: Rx discard flush timeout
While configuring the uart buadrate using set_termios
function, timeout is occuring for rx discard flush
operation. This is due to the driver is waiting for
less time for the DMA call back to occur after programing
the DMA register with the discard flush command.

To overcome this issue increased the timeout to 300ms.

[backport from 3.4 kernel commit 067d1ab83194db80e97904fa090af8f9322449fd]

Change-Id: I3876b2539530a2417fb2ad70711c0e5cbde3670f
Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>