dts: ipq806x: Remove rootfs and scratch partitions
By default, the rootfs and scratch partitions are defined for
spi driver in dts file. So these partitions are created in NAND
boot while spi driver probes. These paritions are not required.
Since the partitions will be created in SBL and passed into SMEM
which will be patched into dtb as mtd partitions in u-boot.

Change-Id: I4fbc38e484bb6f9d0a4469fede605fbb5cfc6425
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>