bridge: inherit slave devices needed_headroom
cherry-picked from linux-stable:
https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/?id=fd094808a06e290432fc13c09aae808aea34d2ca

Some slave devices may have set a dev->needed_headroom value which is
different than the default one, most likely in order to prepend a
hardware descriptor in front of the Ethernet frame to send. Whenever a
new slave is added to a bridge, ensure that we update the
needed_headroom value accordingly to account for the slave
needed_headroom value.

Change-Id: I3a18fa1ef7180613706589891ba8a0a2e1462bbe
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Stephen Wang <wstephen@codeaurora.org>