ipq806x: dts: Remove rootfs and scratch partitions
By default, the rootfs and scratch partitions are defined
for spi driver in dts file. So these partitions are
created in NAND boot while spi driver probes. These
partitions are not required. Since the partitions will be
created in SBL and passed into SMEM, which will be patched
into dtb as mtd partitions in u-boot.

Change-Id: Icf3e13641fdade35634fb660adad7ea4b4f476b1
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>