USB: fix invalid memory access in hub_activate()
Commit 8520f38099cc ("USB: change hub initialization sleeps to
delayed_work") changed the hub_activate() routine to make part of it
run in a workqueue.  However, the commit failed to take a reference to
the usb_hub structure or to lock the hub interface while doing so.  As
a result, if a hub is plugged in and quickly unplugged before the work
routine can run, the routine will try to access memory that has been
deallocated.  Or, if the hub is unplugged while the routine is
running, the memory may be deallocated while it is in active use.

This patch fixes the problem by taking a reference to the usb_hub at
the start of hub_activate() and releasing it at the end (when the work
is finished), and by locking the hub interface while the work routine
is running. It also adds a check at the start of the routine to see if
the hub has already been disconnected, in which nothing should be done.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Reported-by: Alexandru Cornea <alexandru.cornea@intel.com>
Tested-by: Alexandru Cornea <alexandru.cornea@intel.com>
Fixes: 8520f38099cc ("USB: change hub initialization sleeps to delayed_work")
CC: <stable@vger.kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Git-commit: e50293ef9775c5f1cf3fcc093037dd6a8c5684ea
Git-repo:  https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git
Signed-off-by: Ravi Kumar Siddojigari <rsiddoji@codeaurora.org>
(cherry picked from commit 4a38540855ec866d6b1e896f21fa57ab88fdfe72)

Change-Id: I92f75e14712db6cb09a86c521b677e76faa4416b
Signed-off-by: Sanjeev N <nsanjeev@codeaurora.org>