CHROMIUM: mac80211: send packets to mesh peers without routing
Mesh links that are not picked by HWMP will not be
exercised and therefore we can't get link quality
updates from such links. To exercise these links
we need a way to bypass mesh routing, but only for
the data packets we inject, not for normal packets.

The assumption is that the injected packets will
have DA = all 0s and SA = mesh peer MAC. The reason
is that mac80211 mesh code is only interested in DA
and by putting 0 pattern in DA we can eliminate extra
logic for normal packets. In the exception case, if
we see 0s then correct the addrs, put the mesh
peer in RA/DA/meshDA and bypass HWMP.

There is a compile flag to turn on/off this.

BUG=b:26546424
TEST=manually on 4 WWs and 1 AK forming a mesh

Reviewed-on: https://chromium-review.googlesource.com/322775
Commit-Ready: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>

Change-Id: I49a8e1ae87281fa275343daf544f7057f072dddc
Signed-off-by: Julan Hsu <julanhsu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/340405
Commit-Ready: Yixiang Li <yixiang@google.com>
Tested-by: Yixiang Li <yixiang@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>