ipq806x: mmc: Modified signal voltage switching for UHS cards
This patch updates the signal voltage switching procedure
for UHS cards. SD clock needs to be stopped before switching
the voltage by toggling UHS gpio pin. And the clock will be
re-enabled after the UHS gpio pin is toggled. This patch solves
the issue of certain cards not getting detected in UHS speed.

Change-Id: Ifd2ee0376c5c51b6ef3baddce25c164d1005742f
Signed-off-by: Vasudevan Murugesan <vmuruges@codeaurora.org>