[ipq40xx][edma] Move phy_connect code before register_netdevice
This is done to ensure stabilty while doing a insmod/rmmod of edma
module. Once in a while, we observe that while doing a rmmod/insmod,
our netdevice registers first and then we get hold of a phydev.

Since, out netdevice is already registered, there is a chance of edma_open
getting called with our phydev pointers in inconsistent state.

So, this patch avoids that scenario by first doing a phy_connect and then doing
a register_netdevice.

genphy_resume is not required now before phy_start, the fix is implemented in
the kernel phy code itself.

Change-Id: Iaa9c733211fa322373217d438cd1af946e04b829
Signed-off-by: Poddar,Sourav <souravp@codeaurora.org>