skbuff_debug: don't update free stack until after we confirm this is deactive
If we update the stack, we just see the second free stack (another
victim). With this change we won't update the free stack until after we
have confurmed the state of the SKB.

You will still have the free stack for the second free too in the
backtrace of the BUG() output.

Change-Id: I94271f3e03786437c07b23f42c0dcb0591a88836
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>