ipq806x: Added support for reading MAC address from eMMC
This patch seperates the nss gmac fixups in a new function
nss_gmac_fixup_pdata. Also modifies the function nss_fixup_platform_data
to have only the generic nss fixups for pdata. The function
nss_gmac_fixup_pdata is made as extern, So that qca-nss-gmac driver
can call during the module_init. This  approach is needed due to
the reason that eMMC block layer gets initialized after all the
__init calls completed. Due to this reading of eMMC ART partition
 is not possible within the function nss_fixup_platform_data.
Also it is more appropriate to call this gmac fixups in
qca-nss-gmac driver's module_init. So that it always make sure that
the initialization required for qca-nss-gmac driver completed within
it's module_init.

Change-Id: I5fdc1c2ad450328a2c4b445affccc66c6b6cb9d1
Signed-off-by: Vasudevan Murugesan <vmuruges@codeaurora.org>