ipq806x: wdt: Added the support for flash crashdump
This patch adds the support for collection of crashdump
information in flash. It copies the required data or
physical address of data to the page allocated for
crashdump cpu context dump. In the case of kernel panic,
uboot will extract this information and will write to
flash memory.

This change provides a reference to add any other
information needed to be saved.

Change-Id: Id2a148f7f8e79d504ed601c1f7f5ef07a71012e0
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>