CHROMIUM: ath10k: Add debugfs interface for per station TPC
This patch will add the support to control the transmit power control
(TPC) of
a station associated.

Following is an example of how you set the TPC for a station

echo 15 >
/sys/kernel/debug/ieee80211/phy0/netdev\:wlan0/stations/<mac_addr_sta>/tpc

Valid values are from 0 to 70 for tpc.

Special value of 0 will revert the fixed tx power to auto tx
power for the station.

Currently, 44 dBm of tx power happens to be the maximum value listed in
regulatory domain database.  for 60 GHz. To be future proof, have the
range
of allowed max value of TPC to 70. Higher values are anyways by lower
layers depending on the regulatory domain.

BUG=chrome-os-partner:46144
TEST=Associate a station. Change the TPC value using above method.
Measure the
TX power used by AP when sending traffic to the station

Change-Id: If787a1efd8933faf3e338569a5521b977b323858
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/339245
Commit-Ready: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>
Tested-by: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>