bridge: Extend Proxy ARP design to allow optional rules for Wi-Fi
This extends the design in commit 958501163ddd ("bridge: Add support for
IEEE 802.11 Proxy ARP") with optional set of rules that are needed to
meet the IEEE 802.11 and Hotspot 2.0 requirements for ProxyARP. The
previously added BR_PROXYARP behavior is left as-is and a new
BR_PROXYARP_WIFI alternative is added so that this behavior can be
configured from user space when required.

In addition, this enables proxyarp functionality for unicast ARP
requests for both BR_PROXYARP and BR_PROXYARP_WIFI since it is possible
to use unicast as well as broadcast for these frames.

The key differences in functionality:

BR_PROXYARP:
 - uses the flag on the bridge port on which the request frame was
   received to determine whether to reply
 - block bridge port flooding completely on ports that enable proxy ARP

BR_PROXYARP_WIFI:
 - uses the flag on the bridge port to which the target device of the
   request belongs
 - block bridge port flooding selectively based on whether the proxyarp
   functionality replied

Change-Id: Ia87d8ac7d32d4d1ee29602187f9b0ff0aefe7104
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>