ipq40xx: nand: Modified the EUCLEAN and EBADMSG error handling
The current ipq40xx driver comes out of read loop if it gets
EBADMSG or EUCLEAN. The UBI layer requires the complete data
of requested length in case of these errors otherwise mounting
will fail.

This patch checks for these errors and store it in ecc_status
and continue the reading of all the requested length of data.
This ecc_status will be checked during return from nand read
function. The return value will be replaced with ecc_status
if it is non zero.

Change-Id: I2147035ca634d4f7aa1f226d548f1c9b8e2a9eb8
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>